Complete migration to PySide6
If anyone ever tells you to start learning GUI's with Tkinter because it's easier, it's a lie. They're lying to you. It's easier to get text to appear, sure, but the second you want a GUI that even remotely resembles a GUI, get ready to write every single little helper you'll need because Tkinter doesn't help you at all.